==> models/saiva-3-day-hosp-v6/src/saiva/notebooks/08-model-config.ipynb <==
### Create trained models in db

import sys
import json

from omegaconf import OmegaConf

## Load config

from saiva.model.shared.constants import saiva_api, LOCAL_TRAINING_CONFIG_PATH
from saiva.training.utils import load_config

config = load_config(LOCAL_TRAINING_CONFIG_PATH)
training_config = config.training_config

with open('model_config.json', 'r') as file:
    model_config = json.load(file)
    
model_config['validation_start_date'] = model_config['valid_start_date']
model_config['validation_end_date'] = model_config['valid_end_date']

training_metadata = training_config.training_metadata
training_metadata['organization_configs'] = [{
    'organization_id': config.organization_id,
    'datasource_id': config.datasource.id

} for config in training_config.organization_configs]

saiva_api.trained_models.create_with_dict(ml_model_org_config_id=None, training_metadata=OmegaConf.to_container(training_metadata, resolve=True), model_config=model_config)
